requirements.txt
numpy>=1.26.0,<2.0.0
pandas==2.2.3
scikit_learn==1.4.2
torch>=2.1.0
torchvision>=0.16.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
supervision==0.25.0
ultralytics==8.3.37
mlflow>=2.9.0
wandb>=0.16.0
tensorboard>=2.15.0
dvc>=3.30.1
dvc[s3]
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.1
pillow>=10.0.0
matplotlib>=3.8.0
pytest>=7.4.0
black>=23.9.1
flake8>=6.1.0
isort>=5.12.0